crypto-bigint: factor `UInt`-related code into modules
Factors `UInt` into submodules per operation/function.

Additionally factors the related tests into those same modules, making
it easier to do TDD for related functionality, and also clearer how each
function is being tested.
