+++ crypto-bigint/src/lib.rs
@@ -41,6 +41,9 @@
 #![forbid(unsafe_code, clippy::unwrap_used)]
 #![warn(missing_docs, rust_2018_idioms, unused_qualifications)]
 
+#[cfg(all(feature = "alloc", test))]
+extern crate alloc;
+
 #[macro_use]
 mod macros;
