+++ crypto-bigint/src/uint/add.rs
@@ -1,0 +1,77 @@
+//! [`UInt`] addition operations.
+
+use super::UInt;
+use crate::limb::{self, Limb};
+use subtle::{ConstantTimeEq, CtOption};
+
+impl<const LIMBS: usize> UInt<LIMBS> {
+    /// Computes `a + b + carry`, returning the result along with the new carry.
+    #[inline(always)]
+    pub const fn adc(&self, rhs: &Self, mut carry: Limb) -> (Self, Limb) {
+        let mut limbs = [0; LIMBS];
+        let mut i = 0;
+
+        while i < LIMBS {
+            let (w, c) = limb::adc(self.limbs[i], rhs.limbs[i], carry);
+            limbs[i] = w;
+            carry = c;
+            i += 1;
+        }
+
+        (Self { limbs }, carry)
+    }
+
+    /// Perform wrapping addition, discarding overflow.
+    pub const fn wrapping_add(&self, rhs: &Self) -> Self {
+        self.adc(rhs, 0).0
+    }
+
+    /// Perform checked addition, returning [`CtOption`] only if the operation
+    /// did not overflow.
+    pub fn checked_add(&self, rhs: &Self) -> CtOption<Self> {
+        let (result, carry) = self.adc(rhs, 0);
+        CtOption::new(result, carry.ct_eq(&0))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::U128;
+
+    #[test]
+    fn adc_no_carry() {
+        let (res, carry) = U128::ZERO.adc(&U128::ONE, 0);
+        assert_eq!(res, U128::ONE);
+        assert_eq!(carry, 0);
+    }
+
+    #[test]
+    fn adc_with_carry() {
+        let (res, carry) = U128::MAX.adc(&U128::ONE, 0);
+
+        assert_eq!(res, U128::ZERO);
+        assert_eq!(carry, 1);
+    }
+
+    #[test]
+    fn wrapping_add_no_carry() {
+        assert_eq!(U128::ZERO.wrapping_add(&U128::ONE), U128::ONE);
+    }
+
+    #[test]
+    fn wrapping_add_with_carry() {
+        assert_eq!(U128::MAX.wrapping_add(&U128::ONE), U128::ZERO);
+    }
+
+    #[test]
+    fn checked_add_ok() {
+        let result = U128::ZERO.checked_add(&U128::ONE);
+        assert_eq!(result.unwrap(), U128::ONE);
+    }
+
+    #[test]
+    fn checked_add_overflow() {
+        let result = U128::MAX.checked_add(&U128::ONE);
+        assert!(!bool::from(result.is_some()));
+    }
+}
