+++ crypto-bigint/src/uint/array.rs
@@ -1,0 +1,169 @@
+//! `generic-array` integration with [`UInt`]
+// TODO(tarcieri): completely phase out `generic-array` when const generics are powerful enough
+
+use super::UInt;
+use crate::{ArrayEncoding, ByteArray, LIMB_BYTES};
+use generic_array::{typenum, GenericArray};
+
+impl<const LIMBS: usize> UInt<LIMBS> {
+    /// Serialize this [`UInt`] as big-endian, writing it into the provided
+    /// byte slice.
+    #[inline]
+    #[cfg_attr(docsrs, doc(cfg(feature = "generic-array")))]
+    pub(crate) fn to_be_bytes(&self, out: &mut [u8]) {
+        debug_assert_eq!(out.len(), LIMB_BYTES * LIMBS);
+
+        for (src, dst) in self
+            .limbs
+            .iter()
+            .rev()
+            .zip(out.chunks_exact_mut(LIMB_BYTES))
+        {
+            dst.copy_from_slice(&src.to_be_bytes());
+        }
+    }
+
+    /// Serialize this [`UInt`] as little-endian, writing it into the provided
+    /// byte slice.
+    #[inline]
+    #[cfg_attr(docsrs, doc(cfg(feature = "generic-array")))]
+    pub(crate) fn to_le_bytes(&self, out: &mut [u8]) {
+        debug_assert_eq!(out.len(), LIMB_BYTES * LIMBS);
+
+        for (src, dst) in self.limbs.iter().zip(out.chunks_exact_mut(LIMB_BYTES)) {
+            dst.copy_from_slice(&src.to_le_bytes());
+        }
+    }
+}
+
+macro_rules! impl_uint_array_encoding {
+    ($(($uint:ident, $bytes:path)),+) => {
+        $(
+            #[cfg_attr(docsrs, doc(cfg(feature = "generic-array")))]
+            impl ArrayEncoding for super::$uint {
+                type ByteSize = $bytes;
+
+                #[inline]
+                fn from_be_byte_array(bytes: &ByteArray<Self>) -> Self {
+                    Self::from_be_bytes(bytes)
+                }
+
+                #[inline]
+                fn from_le_byte_array(bytes: &ByteArray<Self>) -> Self {
+                    Self::from_le_bytes(bytes)
+                }
+
+                #[inline]
+                fn to_be_byte_array(&self) -> ByteArray<Self> {
+                    let mut result = GenericArray::default();
+                    self.to_be_bytes(&mut result);
+                    result
+                }
+
+                #[inline]
+                fn to_le_byte_array(&self) -> ByteArray<Self> {
+                    let mut result = GenericArray::default();
+                    self.to_le_bytes(&mut result);
+                    result
+                }
+            }
+        )+
+     };
+}
+
+// TODO(tarcieri): use `const_evaluatable_checked` when stable to make generic around bits.
+impl_uint_array_encoding! {
+    (U64, typenum::U8),
+    (U128, typenum::U16),
+    (U192, typenum::U24),
+    (U256, typenum::U32),
+    (U384, typenum::U48),
+    (U448, typenum::U56),
+    (U512, typenum::U64),
+    (U768, typenum::U96),
+    (U896, typenum::U112),
+    (U1024, typenum::U128),
+    (U1536, typenum::U192),
+    (U1792, typenum::U224),
+    (U2048, typenum::U256),
+    (U3072, typenum::U384),
+    (U3584, typenum::U448),
+    (U4096, typenum::U512),
+    (U6144, typenum::U768),
+    (U8192, typenum::U1024)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::ArrayEncoding;
+    use hex_literal::hex;
+
+    #[cfg(target_pointer_width = "32")]
+    use crate::U64 as UIntEx;
+
+    #[cfg(target_pointer_width = "64")]
+    use crate::U128 as UIntEx;
+
+    /// Byte array that corresponds to `UIntEx`
+    type ByteArray = crate::ByteArray<UIntEx>;
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_be_byte_array() {
+        let n = UIntEx::from_be_byte_array(&hex!("0011223344556677").into());
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_be_byte_array() {
+        let n = UIntEx::from_be_byte_array(&hex!("00112233445566778899aabbccddeeff").into());
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_le_byte_array() {
+        let n = UIntEx::from_le_byte_array(&hex!("7766554433221100").into());
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_le_byte_array() {
+        let n = UIntEx::from_le_byte_array(&hex!("ffeeddccbbaa99887766554433221100").into());
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn to_be_byte_array() {
+        let expected_bytes = ByteArray::from(hex!("0011223344556677"));
+        let actual_bytes = UIntEx::from_be_byte_array(&expected_bytes).to_be_byte_array();
+        assert_eq!(expected_bytes, actual_bytes);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn to_be_byte_array() {
+        let expected_bytes = ByteArray::from(hex!("00112233445566778899aabbccddeeff"));
+        let actual_bytes = UIntEx::from_be_byte_array(&expected_bytes).to_be_byte_array();
+        assert_eq!(expected_bytes, actual_bytes);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn to_le_byte_array() {
+        let expected_bytes = ByteArray::from(hex!("7766554433221100"));
+        let actual_bytes = UIntEx::from_le_byte_array(&expected_bytes).to_le_byte_array();
+        assert_eq!(expected_bytes, actual_bytes);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn to_le_byte_array() {
+        let expected_bytes = ByteArray::from(hex!("ffeeddccbbaa99887766554433221100"));
+        let actual_bytes = UIntEx::from_le_byte_array(&expected_bytes).to_le_byte_array();
+        assert_eq!(expected_bytes, actual_bytes);
+    }
+}
