+++ crypto-bigint/src/uint/decode.rs
@@ -1,0 +1,253 @@
+//! Const-friendly decoding operations for [`UInt`]
+
+use super::UInt;
+use crate::{Limb, LIMB_BYTES};
+
+impl<const LIMBS: usize> UInt<LIMBS> {
+    /// Create a new [`UInt`] from the provided big endian bytes.
+    pub const fn from_be_bytes(bytes: &[u8]) -> Self {
+        const_assert!(
+            bytes.len() == LIMB_BYTES * LIMBS,
+            "bytes are not the expected size"
+        );
+
+        let mut decoder = Decoder::new();
+        let mut i = 0;
+
+        while i < LIMB_BYTES * LIMBS {
+            i += 1;
+            decoder = decoder.add_byte(bytes[bytes.len() - i]);
+        }
+
+        decoder.finish()
+    }
+
+    /// Create a new [`UInt`] from the provided big endian hex string.
+    pub const fn from_be_hex(hex: &str) -> Self {
+        let bytes = hex.as_bytes();
+
+        const_assert!(
+            bytes.len() == LIMB_BYTES * LIMBS * 2,
+            "hex string is not the expected size"
+        );
+
+        let mut decoder = Decoder::new();
+        let mut i = 0;
+
+        while i < LIMB_BYTES * LIMBS * 2 {
+            i += 2;
+            let offset = bytes.len() - i;
+            let byte = decode_hex_byte([bytes[offset], bytes[offset + 1]]);
+            decoder = decoder.add_byte(byte);
+        }
+
+        decoder.finish()
+    }
+
+    /// Create a new [`UInt`] from the provided little endian bytes.
+    pub const fn from_le_bytes(bytes: &[u8]) -> Self {
+        const_assert!(
+            bytes.len() == LIMB_BYTES * LIMBS,
+            "bytes are not the expected size"
+        );
+
+        let mut decoder = Decoder::new();
+        let mut i = 0;
+
+        while i < LIMB_BYTES * LIMBS {
+            decoder = decoder.add_byte(bytes[i]);
+            i += 1;
+        }
+
+        decoder.finish()
+    }
+
+    /// Create a new [`UInt`] from the provided little endian hex string.
+    pub const fn from_le_hex(hex: &str) -> Self {
+        let bytes = hex.as_bytes();
+
+        const_assert!(
+            bytes.len() == LIMB_BYTES * LIMBS * 2,
+            "bytes are not the expected size"
+        );
+
+        let mut decoder = Decoder::new();
+        let mut i = 0;
+
+        while i < LIMB_BYTES * LIMBS * 2 {
+            let byte = decode_hex_byte([bytes[i], bytes[i + 1]]);
+            decoder = decoder.add_byte(byte);
+            i += 2;
+        }
+
+        decoder.finish()
+    }
+}
+
+#[derive(Clone, Debug)]
+struct Decoder<const LIMBS: usize> {
+    /// Limbs being decoded.
+    ///
+    /// Stored from least significant to most significant.
+    limbs: [Limb; LIMBS],
+
+    /// Current limb being decoded.
+    index: usize,
+
+    /// Total number of bytes consumed.
+    bytes: usize,
+}
+
+impl<const LIMBS: usize> Decoder<LIMBS> {
+    /// Create a new decoder.
+    pub const fn new() -> Self {
+        Self {
+            limbs: [0; LIMBS],
+            index: 0,
+            bytes: 0,
+        }
+    }
+
+    /// Add a byte onto the [`UInt`] being decoded.
+    pub const fn add_byte(mut self, byte: u8) -> Self {
+        if self.bytes == LIMB_BYTES {
+            const_assert!(self.index < LIMBS, "too many bytes in UInt");
+            self.index += 1;
+            self.bytes = 0;
+        }
+
+        self.limbs[self.index] |= (byte as Limb) << (self.bytes * 8);
+        self.bytes += 1;
+        self
+    }
+
+    /// Finish decoding a [`UInt`], returning a decoded value only if we've
+    /// received the expected number of bytes.
+    pub const fn finish(self) -> UInt<LIMBS> {
+        const_assert!(self.index == LIMBS - 1, "decoded UInt is missing limbs");
+        const_assert!(self.bytes == LIMB_BYTES, "decoded UInt is missing bytes");
+        UInt { limbs: self.limbs }
+    }
+}
+
+impl<const LIMBS: usize> Default for Decoder<LIMBS> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+/// Decode a single byte encoded as two hexadecimal characters.
+const fn decode_hex_byte(bytes: [u8; 2]) -> u8 {
+    let mut i = 0;
+    let mut result = 0u8;
+
+    while i < 2 {
+        result <<= 4;
+        result |= match bytes[i] {
+            b @ b'0'..=b'9' => b - b'0',
+            b @ b'a'..=b'f' => 10 + b - b'a',
+            b @ b'A'..=b'F' => 10 + b - b'A',
+            b => {
+                const_assert!(
+                    matches!(b, b'0'..=b'9' | b'a' ..= b'f' | b'A'..=b'F'),
+                    "invalid hex byte"
+                );
+                0
+            }
+        };
+
+        i += 1;
+    }
+
+    result
+}
+
+#[cfg(test)]
+mod tests {
+    use hex_literal::hex;
+
+    #[cfg(feature = "alloc")]
+    use {crate::U128, alloc::format};
+
+    #[cfg(target_pointer_width = "32")]
+    use crate::U64 as UIntEx;
+
+    #[cfg(target_pointer_width = "64")]
+    use crate::U128 as UIntEx;
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_be_bytes() {
+        let bytes = hex!("0011223344556677");
+        let n = UIntEx::from_be_bytes(&bytes);
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_be_bytes() {
+        let bytes = hex!("00112233445566778899aabbccddeeff");
+        let n = UIntEx::from_be_bytes(&bytes);
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_le_bytes() {
+        let bytes = hex!("7766554433221100");
+        let n = UIntEx::from_le_bytes(&bytes);
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_le_bytes() {
+        let bytes = hex!("ffeeddccbbaa99887766554433221100");
+        let n = UIntEx::from_le_bytes(&bytes);
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_be_hex() {
+        let n = UIntEx::from_be_hex("0011223344556677");
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_be_hex() {
+        let n = UIntEx::from_be_hex("00112233445566778899aabbccddeeff");
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "32")]
+    fn from_le_hex() {
+        let n = UIntEx::from_le_hex("7766554433221100");
+        assert_eq!(n.limbs(), &[0x44556677, 0x00112233]);
+    }
+
+    #[test]
+    #[cfg(target_pointer_width = "64")]
+    fn from_le_hex() {
+        let n = UIntEx::from_le_hex("ffeeddccbbaa99887766554433221100");
+        assert_eq!(n.limbs(), &[0x8899aabbccddeeff, 0x0011223344556677]);
+    }
+
+    #[cfg(feature = "alloc")]
+    #[test]
+    fn hex_upper() {
+        let hex = "AAAAAAAABBBBBBBBCCCCCCCCDDDDDDDD";
+        let n = U128::from_be_hex(hex);
+        assert_eq!(hex, format!("{:X}", n));
+    }
+
+    #[cfg(feature = "alloc")]
+    #[test]
+    fn hex_lower() {
+        let hex = "aaaaaaaabbbbbbbbccccccccdddddddd";
+        let n = U128::from_be_hex(hex);
+        assert_eq!(hex, format!("{:x}", n));
+    }
+}
