--- crypto-bigint/tests/lib.rs
@@ -1,3 +1,0 @@
-//! Integration tests
-
-mod uint;
